requirements.txt
# This file was generated using teh command: pipreqs .(you can also use pip list)
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
psycopg==3.1.9
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
django-storages[google]
Pillow